+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-21T03:12:18.611Z")
+    created: Optional[datetime] = parse_datetime("2026-01-21T10:02:31.078Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")