+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-21T03:12:14.981Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-21T10:02:27.374Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")