+++ packages/orq-rc/RELEASES.md
@@ -4909,3 +4909,13 @@ Based on:
 - [python v4.2.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.53 - packages/orq-rc
+
+## 2026-01-21 10:05:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.691.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.54] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.54 - packages/orq-rc